Build all PCM plugins referenced by conf files
The libasound package installs many .conf files to support sound cards
and PCM devices, and several of them reference plugins that are not
compiled at build time.

Since the spec file is using a whitelist for the plugins, we have to
make sure all referenced plugins are built.

This commit fixes issues with Pulseaudio detecting HDMI support and
general issues with the alsa-utils package (see TIVI-1021).

Change-Id: I85b84791f73cfe18543ff46b1f002cb462d0ab80